modify header method
